Check for pid of uwsgi

+++ scripts/stop_server.sh
@@ -1,1 +1,5 @@
-killall uwsgi
+#!/usr/bin/env bash
+pidof uwsgi  >/dev/null
+if [[ $? -ne 0 ]]; then
+  killall uwsgi
+fi
